GWCA has now converted to a new communication method: Named Pipes. Thanks a lot to Cronos for his work!
Everything should've been done in a painless way to secure backwards compatibility BUT remember NOT to register WndCallback anymore - it doesn't exist now!

The two experimental functions CmdEx() and CmdCBEx() where removed as of this release. They no longer served a purpose.

All you need to do to use GWCA now, is to include the constants file in the script and inject the Dll!